Improve ersky9x interaction latency
All changes are related to latency improvement when using
the multi-protocol module in an ersky9x remote that is
featuring a special firmware as well.

Changes:
- increase baudrate to 250000 (from 100000). 3ms -> 1.2ms tx time
- tweak scheduler to call Update_All() more often so that
  update_serial_data() is called right before tx packet go out
- preemptively call frskyX_data_frame() to prepare outgoing frame
  buffer with the most up to date data as it comes in.
- sync telemetry updates to the packet transmission transmission
  event so the remote can use telemetry updates as a heartbeat
  and sent channel updates ~1ms before packets are sent over the air.

Those change require the ersky9x remote to be flashed
with a custom firmware that syncs the telemetry updates
sent to the module to the tranmission of the channel update
packets over the air.